sql: rework StringAny/JsonbAny coercion
The SQL function selection code previously allowed parameter types to be
declared as "StringAny" or "JsonbAny", representing any type that could
be converted to a string or jsonb, respectively. The function selection
infrastructure was responsible for installing the conversion to string
or jsonb.

This turns out not to generalize well. In PostgreSQL, only the "Any"
psuedotype exists, and the function's *implementation*, rather than the
function selection code, is responsible for casting the arguments to the
correct type, if required. `concat` applies a nonstandard bool -> string
conversion, for example, and the forthcoming `pg_typeof` actually
performs no type conversion at all.

This commit makes our function selection infrastructure work more like
PostgreSQL's. Rather than trying to capture all possible coercions in
ParamType, we include only the "Any" type in ParamType, and let the
few functions that take "Any" arguments apply whatever bespoke
conversions they want. The big win is that the cast and coercion
routines no longer need to think about "JsonAny" as a concept, as that
complexity is entirely isolated in the "typeconv::to_jsonb" function.