sql-parser: standardize WITH options parsing
There are currently five ways `WITH` is used in our grammar when configuring options:

1. `WITH SNAPSHOT`
2. `WITHOUT SNAPSHOT`
3. `WITH (option value)`
4. `WITH (option = value)`
5. `WITH (option)`

We want to reduce this to some more standard form. Postgres has historic
support for `WITHOUT` but does not appear to have used it since, and
so we would like to remove that keyword. `WITH KEYWORD` doesn't support
multiple options or arguments to them, so we would also like to remove
it. Both of these have been replaced with 4 and 5 from the above list.

`COPY` is a bit special in that we would like to preserve its backward
compat with most existing Postgres applications, so it is allowed to
omit the `=` between option names and values. `COPY` has no user-facing
changes due to this.

`TAIL` has thus had a breaking syntax change. Due to all of the recent
churn with `TAIL` we think it's ok to not preserve backward compat in
its syntax.

Other statements that use `WITH` and `WITHOUT` will be changed in
the future.