sql: tweak statement description interface
This commit tweaks the output of sql::plan::describe so that
StatementDesc is the sole authority on the output of the statement.
Previously StatementDesc contained parameter types as a Vec<ScalarType>,
but sql::plan::describe also returned parameter types as a
Vec<pgrepr::Type> alongside the StatementDesc.

This commit also changes the `send: bool` field to `is_copy: bool`,
where `send == !is_copy`. The change seems minor, but I think it more
clearly separates the responsibilities; the sql crate doesn't really
want to determine whether the pgwire crate sends back a RowDescription
to the client or not. Instead the sql crate just surfaces some
information ("was this a COPY statement?") that the pgwire crate can use
as it likes. Some additional pgtest coverage is added to account for
this change.