sql-parser: get out of the error rendering business
PostgreSQL errors have a facility for reporting the character offset
associated with an error reported to the client.  The psql command line
program, and possibly others, use this information to output a visual
marker of where the error occured in the query:

    $ psql -c 'select 1 order wrong'
    ERROR:  syntax error at or near "wrong"
    LINE 1: select 1 order wrong
                           ^

Materialize keeps track of the same information in the SQL parser, but
was not reporting this information to the client directly. Instead, it
was rendering its own visual error message like so:

    $ psql -h localhost -p 6875 materialize -c 'select 1 order wrong'
    ERROR:  Parse error:
    select 1 order wrong
             ^^^^^
    Expected end of statement, found: order

This approach has several downsides:

    * Rendering the error caret is not trivial in the face of unicode
      characters. psql gets this right, but Materialize does not.

    * Rendering the error caret correctly in general is impossible,
      since e.g. other client drivers may render the error message in a
      non-fixed-width font. psql can assume terminal output, but the
      fmt::Display implementation for ParserError should not.

    * Adding the structured "position" field to the error response
      causes psql to render its own error cursor beneath Materialize's.

So, this commit removes our custom error rendering code, and instead
passes the correct information to the client so e.g. psql can do its own
error rendering. Various cleanups result from this work:

    * Since PostgreSQL only supports associating a single character
      offset with an error, rather than a range, various token position
      tracking code gets much simpler.

    * Parser errors no longer need to hold a reference to the input
      string, so the SQL parser can take an `&str` instead of a `String`
      as input, and many of its callers can avoid clones.

The datadriven tests for the SQL parser continue to roll their own caret
rendering, for ease of visually inspecting the test files. But this code
has been simplified, since we now render only a single character rather
than a range, and taught to correctly compute unicode widths. Further
bugs in this code are very low priority, since they only impact our
internal tests.

While some error messages do get a bit worse, since only a single
character is highlighted instead of a whole range, I think that is a
minor price to pay to fix the issues listed above.