sql: fix crash with zero-arity tables
These are technically allowed, though not with the syntax we were
permitting. Drop support for

    CREATE TABLE t

but, like PostgreSQL, continue to support:

    CREATE TABLE t ()

Also, fix the crash that causes

    INSERT INTO t VALUES (1)

to panic rather than producing a proper error message.

While I'm here, disallow trailing commas in column specifications in
CREATE TABLE. This is historical baggage from upstream. We want to
follow PostgreSQL more strictly, which does not permit trailing commas
in this context.

Fix #4759.