sql-parser: handle subquery nesting like postgres
PostgreSQL has a number of grammatical quirks around subqueries.
WITH, ORDER BY, and LIMIT can "escape" top-level nesting, as in:

    (SELECT 1 ORDER BY 1)
    (SELECT 1 LIMIT 1)
    (WITH a AS (SELECT 1) SELECT 2)

And therefore these queries must be rejected with "multiple <blah>
clauses not allowed":

    (SELECT 1 ORDER BY 1) ORDER BY 1
    (SELECT 1 LIMIT 1) LIMIT 1
    WITH a AS (SELECT 1) (WITH b AS (SELECT 2) SELECT 3)

But PostgreSQL also correctly handles corner cases of the grammar, where
nested SELECTs contain both extra parentheses and set operations, as in:

    SELECT (((SELECT 2)) UNION SELECT 2)

This turns out to be tricky to handle as the grammar is quite ambiguous.

This commit makes all of the cases listed above match PostgreSQL, and
enables several SLTs that were disabled to exercise the new behavior.