periodically flush source_info metrics
Updates to the source_info are written once we are caught up with the
source to avoid incurring overhead on every source message. However,
during bootstrap it can take a long time until we're caught up with the
source, and that's exactly when a customer likely wants to view
progress.

Source will now flush source_info metrics every N messages to still show
progress in such cases.