repr, pgrepr, sql-parser, func: support map types (#4801)
Adds internal infrastructure to ingest map data. Specifically
adds support for ingesting casted strings as maps. For example,

    '{a=>1}'::map[text=>int]

Can now be parsed and internally represented as a Datum::Map with a
single (a,1) key, value pair. Map keys must be of type text, map values
can be of any non-nested scalar type. (Please note that nested maps
are explicitly not supported by this change.)

This change creates a hardcoded OID for the map type, which is not
a standard Postgres type, similar to the OID created for the list type.