Reduce the number of metrics we send for tracking "in maintenance"
We added the `mz_arrangement_maintenance_active_info` metric to track whether
any specific arrangement is spending all of its time in maintenance.
Unfortunately that means that it grows with the number of arrangements,
possibly to hundreds or thousands of metrics.

With this new metric we will only see one metric per worker. It is still
possible to determine which arrangement is spending all its time in mainenance
by looking at the `mz_arrangement_maintenance_seconds_total` metric, which will
show the proportion of time taken up by each arrangement's maintenance.