sql: privatize more of the typeconv api
The only caller of typeconv::get_direct_cast is only interested in
whether the cast exists, not getting a hold of the actual cast
operation. So privatize CastOp and get_direct_cast, and replace them
with a can_cast function. This gives us more flexibility to rework the
API internally.