sql: jam recursive casts into the global cast map
Casting lists and map types is tricky, because the casts must proceed
recursively on the element types. So casts for these types were
previously handled separately from the other types of casts.  A small
bug resulted from this design, as the "can_cast" function was unaware of
these special list and map casts.

Restructure the cast map so that it can describe the recursive casts
necessary for maps and lists. The gist is that the cast map now stores
"cast templates" which can describe a suite of casts, like "string to
any list", rather than just individual casts between two concrete types.