coord,sql: combine resolution/lookup in catalog API
Previously, interacting with the catalog required an annoying bit of
boilerplate:

     let name = scx.catalog.resolve_item(name)?;
     let item = scx.catalog.get_item(&name);

The first call "resolved" a partially-specified item name to a full
name; the second call then looked up the resolved full name in the
catalog to pull out the actual catalog item. Simplify the API to:

     let item = scx.catalog.resolve_item(name)?;

Code that needs access to the fully-specified name can pull it out of
the item on demand via `item.name()`.

The same change is applied to `Catalog::resolve_database` and
`Catalog::resolve_schema`. These methods now return a database and
schema object, respectively, to mirror the item object returned by
Catalog::resolve_item, and are much less awkward to use than before.

There should be no behavioral changes in this PR.