pgwire: add support for FETCH timeouts
Implement this with our extension to the FETCH syntax which now supports
a timeout option. If the requested number of rows has not been returned
within the requested timeout, complete the statement anyway. This allows
drivers that buffer all results to get their data within definable time
and memory bounds.

The default timeout is 0s. This will return up to the requested number
of rows as long as they are immediately available.