sql: minor adjustments to pgcrypto support
Move the pgcrypto SLTs into their own files that mirror the upstream
sources. This makes it easier to integrate upstream changes and gives us
a place to install the appropriate license headers, since technically
these tests are licensed under the PostgreSQL license, not our usual
license. (These tests are probably simple enough not to be
copyrightable, but best to be rigorous.)

Also fix a buglet in the type signature of the digest function. To match
pgcrypto, either both data and key need to have type `text` or both need
to have type `bytea`.

Finally fix a typo in the function documentation so that the description
of the function renders properly in HTML.