Add workflow to restore Kafka from snapshots (#4929)
These workflows assume we only care about key, value, and timestamp for
each message in Kafka. For each message in a topic, it preserves the
original value of these in an Apache Arrow file. Arrow was chosen as a
handy method for serializing / deserializing data that also works with
many other tools in the Python ecosystem.

There are two workflows, one to run chbench and capture a snapshot. The
other workflow is to start a kafka cluster, restore the topic contents
and then measure how long Materialize takes to generate the same results
as measured by the snapshot.