Set chbench Kafka log retention to 1 year (#5064)
The default is 7 days and breaks the snapshot replay benchmark in 3
interesting ways. First, it can cause the test to hang because
Materialize fails to read all messages before they are deleted. Second,
it causes Materialize to fail to generate the same answers because it
fails to read all messages, even if it gets to the end of the topics.
And finally, it prevents Materialize from reading messages at all after
being restarted, as all messages have been removed.

Fixes #5059
Fixes #5030
Fixes #5024