pgwire: change FETCH with TAIL to wait for rows if there are none
Previously FETCH with TAIL would return immediately if there were no
rows available. While useful, this is generally not what users need.
Change FETCH to default to waiting for some rows to be available and
return those. This allows users to call FETCH in a loop without
worrying about backoffs or timeouts and always get available rows as
soon as possible.

Add an ALL option to FETCH to return ALL available rows.