Example application of the Wikipedia Demo (#5021)
This application is a demonstration showing how to build an event-driven
data pipeline all the way to the browser. The example application comes
from our Getting Started Guide, where materialized is configured with
three materialized views that update as new edits are appended to the
`wikirecent` source.

A Python web application serves a basic page that renders
both the current value for `counter` and a visualization of the `top10`
table. These counts and visualization are updated as the underlying
materialized views are updated, using the TAIL command.