catalog-compat: test multiple golden versions
Testing with multiple "golden" catalog versions makes it possible to
test migration of features added in newer Materialize versions, like
tables.

In service of this goal, teach testdrive the "--no-reset" option, which
avoids resetting Materialize state at the beginning of the testdrive
script. This makes using testdrive in the catalog-compat script much
simpler.