downgrade capabilities before reading from a source
Today we only downgrade capabilities after reading from a source and
after timestamp_frequency_ms has passed. For BYO this doesn't affect
anything, but for realtime this means that our initial portion of data
will be timestamped as '1'.

This can be particularly confusing if someone ingests a source that is
no longer being updated and the entire ingest occurs within
timestamp_frequency_ms - in that case mz_source_info will show timestamp
1 in perpetuity (at least one customer has hit this). It is technically
  correct, 1 is the logical timestamp that we used and there are no
functional problems from this behavior.  But it makes for confusing
logging and it's not in the spirit of how realtime is supposed to work.

To work around this we construct ConsistencyInfo such that it is
initially stale and attempt to downgrade capabilities before reading
from a source.