dataflow: swap notify for inotify crate
The notify crate is largely unmaintained, and trying to provide a
consistent filesystem notification API across platforms is a fool's
errand, in my opinion. We only particularly care about performance on
Linux, so switch to using the inotify crate, which is well maintained.
Other platforms now use the 100ms polling fallback that has been working
well on macOS.

This eliminates a number of long-standing duplicate dependencies
associated with Mio v0.6/Tokio v0.2, most importantly the long-deprected
net2 and socket crates.