ore: use vendored openssl
The last commit (fa0e132) introduced a dependency on tokio_openssl in
ore, which transitively introduces a dependency on openssl where it
didn't exist before. So we need to enable the same "vendored" feature
that we use elsewhere to build a bundled copy of openssl, rather than
linking against the system copy.

This only applies when running the unit tests in ore. When building the
materialized binary, for example, openssl's "vendored" feature is
enabled by the materialized crate's Cargo.toml.