coord: refactor serve loop to be safer
In a few places some of the messages handled by the coord serve loop
would "return", expecting that this would go on to handle the next
message. However return here exited the entire coord loop causing
materialize to exit completely.

In order to prevent us from making this mistake again, refactor the
serve loop so that each message type is handled in its own function. This
does the correct thing for return (stop the current message processing
and proceed), which is sometimes needed in error paths. The serve loop
is now easily scannable by the eye and it's much easier to understand
what's going on.