testdrive: simplify and comment consolidation test script
I was having trouble understanding what the consolidation.td test script
was meant to test. This commit simplifies that script where possible and
leaves explanatory comments throughout.

It's probably easiest to review this change as if it were new code. The
changes were substantial enough that the diff isn't very helpful.