materialized: use structopt for main binary
Convert the materialized binary to use structopt. This is a follow-up to
PR #5298, which did the same for all of our utility binaries. The idea
is to have just one option parsing solution (i.e., structopt) throughout
our codebase.

The big benefit of using structopt is that it has built-in support for
reading flag values out of environment variables, which fixes #4191.