Use new cursor.stream method for wikirecent demo (#5422)
This method maps naturally to our usage of TAIL in this scenario, so
let's use it. DECLARE/FETCH is useful if we wanted to iterate a subset
of rows, issue a FETCH with timeout or have multiple cursors on a single
connection. In this case, we just want an endless list of rows, so
stream is the perfect API choice.